writing session status file done
